// dpr/types/api.d.ts
export interface components {
  schemas: {
// ...
    StatementExecutionStatus: {
      /** @description The status of the statement execution. */
      status: string
      /**
       * Format: int64
       * @description The amount of time in nanoseconds that the statement ran.
       * @example 10562762848
       */
      duration: number
      /**
       * Format: int64
       * @description The number of rows returned from the query.
       * @example 10
       */
      resultRows: number
      /**
       * Format: int64
       * @description The size in bytes of the returned results. A -1 indicates the value is null.
       * @example 0
       */
      resultSize?: number
      /** @description Contains a short description of the error that occurred. */
      error?: string
      /**
       * Format: int32
       * @description Specific to Athena queries. An integer value that specifies the category of a query failure error. The following list shows the category for each integer value.
       *     1 - System
       *     2 - User
       *     3 - Other
       */
      errorCategory?: number
      /** @description Specific to Athena queries. Further detail about the status of the query. */
      stateChangeReason?: string
    }
// ...
  }
  responses: never
  parameters: never
  requestBodies: never
  headers: never
  pathItems: never
}